refactor: Add RecordSerbice
